requirements.txt
mysql-connector-python==2.0.4
wheel==0.24.0
